Removed mxContext from core.
