+++ src/gluonts/model/predictor.py
@@ -45,7 +45,6 @@ from gluonts.core.component import (
     DType,
     equals,
     from_hyperparameters,
-    get_mxnet_context,
     validated,
 )
 from gluonts.core.exception import GluonTSException
@@ -53,6 +52,7 @@ from gluonts.core.serde import dump_json, fqname_for, load_json
 from gluonts.dataset.common import DataEntry, Dataset, ListDataset
 from gluonts.dataset.loader import DataBatch, InferenceDataLoader
 from gluonts.model.forecast import Forecast
+from gluonts.mx.context import get_mxnet_context
 from gluonts.mx.distribution import Distribution, DistributionOutput
 from gluonts.support.util import (
     export_repr_block,
