+++ src/gluonts/mx/representation/global_relative_binning.py
@@ -20,9 +20,10 @@ import mxnet as mx
 import numpy as np
 
 # First-party imports
-from gluonts.core.component import get_mxnet_context, validated
+from gluonts.core.component import validated
 from gluonts.dataset.common import Dataset
 from gluonts.model.common import Tensor
+from gluonts.mx.context import get_mxnet_context
 
 from .binning_helpers import (
     bin_edges_from_bin_centers,
