+++ src/gluonts/mx/representation/hybrid_representation.py
@@ -19,9 +19,10 @@ import mxnet as mx
 import numpy as np
 
 # First-party imports
-from gluonts.core.component import get_mxnet_context, validated
+from gluonts.core.component import validated
 from gluonts.dataset.common import Dataset
 from gluonts.model.common import Tensor
+from gluonts.mx.context import get_mxnet_context
 
 from .representation import Representation
