+++ src/gluonts/mx/representation/representation.py
@@ -21,9 +21,10 @@ import numpy as np
 from mxnet.gluon import nn
 
 # First-party imports
-from gluonts.core.component import get_mxnet_context, validated
+from gluonts.core.component import validated
 from gluonts.dataset.common import Dataset
 from gluonts.model.common import Tensor
+from gluonts.mx.context import get_mxnet_context
 
 
 class Representation(nn.HybridBlock):
